Pull up PropagateNullable
