+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/optimizer/PropagateNullable.java
@@ -10,7 +10,7 @@ package org.elasticsearch.xpack.esql.optimizer;
 import org.elasticsearch.xpack.esql.core.expression.Expression;
 import org.elasticsearch.xpack.esql.core.expression.predicate.logical.And;
 
-public class PropagateNullable extends LogicalPlanOptimizer.PropagateNullable {
+public class PropagateNullable extends org.elasticsearch.xpack.esql.optimizer.rules.PropagateNullable {
     @Override
     public Expression rule(And and) {
         return super.rule(and);
